Update mailer test, linting, and update packages for vulnerabilities
