+++ Gemfile
@@ -6,7 +6,7 @@ ruby "3.2.2"
 gem "rails", "~> 7.1.1"
 gem "sprockets-rails"
 gem "pg", "~> 1.1"
-gem "puma", "~> 6.0"
+gem "puma", "~> 6.4.0"
 gem "cssbundling-rails", "~> 1.3", ">= 1.3.3"
 gem "jsbundling-rails"
 gem "turbo-rails"
