+++ app/services/weekend_event_creator.rb
